调整 core 模块的文件结构
