--- metasequoia_sql/core/node/statement_set.py
@@ -1,27 +1,0 @@
-"""
-抽象语法树（AST）节点：SET 语句
-"""
-
-import dataclasses
-
-from metasequoia_sql.core.node.abc_node import ASTStatementBase
-from metasequoia_sql.core.sql_type import SQLType
-
-__all__ = [
-    "ASTSetStatement",  # SET 语句
-]
-
-
-# ---------------------------------------- SET 语句 ----------------------------------------
-
-
-@dataclasses.dataclass(slots=True, frozen=True, eq=True)
-class ASTSetStatement(ASTStatementBase):
-    """SQL 语句"""
-
-    config_name: str = dataclasses.field(kw_only=True)
-    config_value: str = dataclasses.field(kw_only=True)
-
-    def source(self, sql_type: SQLType = SQLType.DEFAULT) -> str:
-        """返回语法节点的 SQL 源码"""
-        return f"SET {self.config_name} = {self.config_value}"
